[KFI]refactor(document-viewer): created comment model + rename
