+++ packages/sn-document-viewer-react/src/components/index.ts
@@ -8,4 +8,4 @@ export { PageList } from './PageList'
 export * from './document-widgets'
 export * from './page-widgets'
 export * from './styles'
-export { Comment } from './Comment'
+export { CommentComponent } from './Comment'
