+++ packages/sn-document-viewer-react/src/models/index.ts
@@ -4,3 +4,4 @@ export * from './PageAttribute'
 export * from './PreviewImageData'
 export * from './Shapes'
 export * from './Theming'
+export * from './Comment'
